pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "polymarket-researcher"
version = "0.1.0"
description = "Oracle-impulse strategy collector, backtester, and executor for Polymarket"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.2",
    "websockets>=14",
    "httpx>=0.28",
    "pandas>=2.2",
    "pyarrow>=19",
    "pydantic>=2.10",
    "python-dotenv>=1.0",
    "rich>=13",
    "py-clob-client>=0.18",
    "aiohttp>=3.11",
]

[project.optional-dependencies]
ml = [
    "xgboost>=2.1",
    "scikit-learn>=1.6",
    "optuna>=4.1",
    "plotly>=5.24",
]
rl = [
    "torch>=2.5",
    "stable-baselines3>=2.4",
]

[project.scripts]
collect = "collect:main"

[tool.hatch.build.targets.wheel]
packages = ["collector", "db", "backtest", "ml", "executor", "paper_trader"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
